o  fixed references to ints and related updates.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497971 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-webapp/src/main/java/org/apache/maven/continuum/web/action/AddProjectAction.java
@@ -62,7 +62,7 @@ public class AddProjectAction
 
     private Collection projectGroups;
 
-    private int selectedProjectGroup;
+    private long selectedProjectGroup;
 
     public void validate()
     {
@@ -229,12 +229,12 @@ public class AddProjectAction
         this.projectGroups = projectGroups;
     }
 
-    public int getSelectedProjectGroup()
+    public long getSelectedProjectGroup()
     {
         return selectedProjectGroup;
     }
 
-    public void setSelectedProjectGroup( int selectedProjectGroup )
+    public void setSelectedProjectGroup( long selectedProjectGroup )
     {
         this.selectedProjectGroup = selectedProjectGroup;
     }
